+++ types/ethereum.go
@@ -97,9 +97,10 @@ func (c *EthereumConfig) VestingBridge() EthereumContract {
 	return c.vestingBridge
 }
 
-// StakingBridgeAddresses returns the staking bridge and/or the token vesting
-// contract address as both can be used to get information needed by the staking
-// process.
+// StakingBridgeAddresses returns the registered staking bridge addresses. It
+// might return the staking bridge, or the token vesting, or both contract
+// address. The vesting contract can also be used to get information needed by
+// the staking engine.
 func (c *EthereumConfig) StakingBridgeAddresses() []ethcmn.Address {
 	var addresses []ethcmn.Address
 
@@ -113,9 +114,30 @@ func (c *EthereumConfig) StakingBridgeAddresses() []ethcmn.Address {
 	return addresses
 }
 
+// StakingBridges returns the registered staking bridges. It might
+// return the staking bridge, or the token vesting, or both contract address.
+// The vesting contract can also be used to get information needed by the
+// staking engine.
+func (c *EthereumConfig) StakingBridges() []EthereumContract {
+	var bridges []EthereumContract
+
+	if c.stakingBridge.HasAddress() {
+		bridges = append(bridges, c.stakingBridge)
+	}
+	if c.vestingBridge.HasAddress() {
+		bridges = append(bridges, c.vestingBridge)
+	}
+
+	return bridges
+}
+
 type EthereumContract struct {
 	address               string
 	deploymentBlockHeight uint64
+}
+
+func (c EthereumContract) DeploymentBlockHeight() uint64 {
+	return c.deploymentBlockHeight
 }
 
 func (c EthereumContract) HasAddress() bool {